refactor: moved transaction list utility functions into a separate file
